Refactor aws client
Added an interface Client, that abstracts AWS client calls. This
indirection helps using gomock to create tests. See ec2_test.go as an
example for this.

As of this commit only the client usage of ec2.go and parts of
route53.go have been updated. The rest of the codebase still requires
this refactoring.